inverstment/requirements.txt
django
djangorestframework
psycopg2-binary
python manage.py createsuperuser

